[Feat] 북마크 눌렀을 때, 저장 구현 세부사항 - 진행 중
